refactor: move data source options
