+++ crates/sail-data-source/build.rs
@@ -7,8 +7,6 @@ fn main() -> Result<(), Box<dyn std::error::Error>> {
     build_data_source_options("Json", "json")?;
     build_data_source_options("Parquet", "parquet")?;
     build_data_source_options("Text", "text")?;
-    build_data_source_options("Delta", "delta")?;
-    build_data_source_options("Iceberg", "iceberg")?;
     build_data_source_options("Socket", "socket")?;
     build_data_source_options("Rate", "rate")?;
     build_data_source_options("Console", "console")?;
