+++ crates/sail-data-source/data/options/console.yaml
