+++ crates/sail-data-source/data/options/parquet.yaml
