--- crates/sail-data-source/src/formats/delta/mod.rs
@@ -1,1 +1,0 @@
-pub mod options;
