--- crates/sail-data-source/src/formats/delta/options.rs
@@ -1,25 +1,0 @@
-use datafusion_session::Session;
-use sail_common_datafusion::datasource::OptionLayer;
-
-use crate::error::DataSourceResult;
-use crate::options::{gen, BuildPartialOptions, PartialOptions, ResolveOptions};
-
-impl ResolveOptions for gen::DeltaReadOptions {
-    fn resolve(_ctx: &dyn Session, options: Vec<OptionLayer>) -> DataSourceResult<Self> {
-        let mut partial = gen::DeltaReadPartialOptions::initialize();
-        for layer in options {
-            partial.merge(layer.build_partial_options()?);
-        }
-        partial.finalize()
-    }
-}
-
-impl ResolveOptions for gen::DeltaWriteOptions {
-    fn resolve(_ctx: &dyn Session, options: Vec<OptionLayer>) -> DataSourceResult<Self> {
-        let mut partial = gen::DeltaWritePartialOptions::initialize();
-        for layer in options {
-            partial.merge(layer.build_partial_options()?);
-        }
-        partial.finalize()
-    }
-}
