--- crates/sail-data-source/src/formats/iceberg/mod.rs
@@ -1,1 +1,0 @@
-pub mod options;
