--- crates/sail-data-source/src/options/types/delta.rs
@@ -1,26 +1,0 @@
-use crate::error::{DataSourceError, DataSourceResult};
-
-#[derive(Debug, Clone, Copy, PartialEq, Eq, Default, serde::Serialize, serde::Deserialize)]
-#[serde(rename_all = "lowercase")]
-pub enum DeltaLogReplayStrategy {
-    #[default]
-    Auto,
-    Sort,
-    Hash,
-}
-
-pub fn parse_delta_log_replay_strategy(
-    key: &str,
-    value: &str,
-) -> DataSourceResult<DeltaLogReplayStrategy> {
-    match value.to_ascii_lowercase().as_str() {
-        "auto" => Ok(DeltaLogReplayStrategy::Auto),
-        "sort" => Ok(DeltaLogReplayStrategy::Sort),
-        "hash" => Ok(DeltaLogReplayStrategy::Hash),
-        _ => Err(DataSourceError::InvalidOption {
-            key: key.to_string(),
-            value: value.to_string(),
-            cause: None,
-        }),
-    }
-}
