--- crates/sail-data-source/src/options/types/mod.rs
@@ -1,1 +1,0 @@
-pub mod delta;
