+++ crates/sail-delta-lake/Cargo.toml
@@ -60,5 +60,8 @@ crc32fast = { workspace = true }
 [dev-dependencies]
 parquet-variant-json = { workspace = true }
 
+[build-dependencies]
+sail-build-scripts = { path = "../sail-build-scripts" }
+
 [lints]
 workspace = true
