+++ crates/sail-delta-lake/src/lib.rs
@@ -15,6 +15,7 @@ pub mod conversion;
 pub mod datasource;
 pub mod deletion_vector;
 mod delta_log;
+pub mod error;
 mod kernel;
 pub mod logical;
 pub mod operations;
