+++ crates/sail-delta-lake/src/physical_plan/writer_options.rs
@@ -13,8 +13,9 @@
 use std::collections::HashMap;
 
 use sail_common_datafusion::catalog::CatalogTableColumnIdentity;
-use sail_data_source::options::gen::DeltaWriteOptions;
 use serde::{Deserialize, Serialize};
+
+use crate::options::gen::DeltaWriteOptions;
 
 /// Options for the Delta Lake writer execution plan.
 ///
