+++ crates/sail-iceberg/Cargo.toml
@@ -55,5 +55,8 @@ rust_decimal = { workspace = true }
 # base64 = { workspace = true }
 murmur3 = { workspace = true }
 
+[build-dependencies]
+sail-build-scripts = { path = "../sail-build-scripts" }
+
 [lints]
 workspace = true
