+++ crates/sail-iceberg/src/error.rs
@@ -1,0 +1,1 @@
+pub use sail_data_source::error::{DataSourceError, DataSourceResult};
