+++ crates/sail-iceberg/src/options.rs
@@ -1,8 +1,26 @@
-use datafusion_session::Session;
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+use datafusion::catalog::Session;
 use sail_common_datafusion::datasource::OptionLayer;
+pub(crate) use sail_data_source::options::{
+    parsers, BuildPartialOptions, PartialOptions, ResolveOptions,
+};
 
 use crate::error::DataSourceResult;
-use crate::options::{gen, BuildPartialOptions, PartialOptions, ResolveOptions};
+
+pub mod gen {
+    include!(concat!(env!("OUT_DIR"), "/options/iceberg.rs"));
+}
 
 impl ResolveOptions for gen::IcebergReadOptions {
     fn resolve(_ctx: &dyn Session, options: Vec<OptionLayer>) -> DataSourceResult<Self> {
