+++ crates/sail-iceberg/src/physical_plan/writer_options.rs
@@ -15,10 +15,10 @@ use std::collections::HashMap;
 use datafusion_common::{DataFusionError, Result};
 use sail_common_datafusion::datasource::OptionLayer;
 use sail_common_datafusion::variant::DEFAULT_VARIANT_INFERENCE_NODE_BUDGET;
-use sail_data_source::options::gen::IcebergWriteOptions;
 use serde::{Deserialize, Serialize};
 
 use crate::operations::write::config::VariantShreddingConfig;
+use crate::options::gen::IcebergWriteOptions;
 
 const PARQUET_SHRED_VARIANTS: &str = "write.parquet.shred-variants";
 const PARQUET_VARIANT_INFERENCE_BUFFER_SIZE: &str = "write.parquet.variant-inference-buffer-size";
